requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Computer Vision
opencv-python==4.8.1.78
pillow==10.0.1
scikit-image==0.21.0

# Deep Learning (CPU-only for deployment)
torch==2.0.1+cpu
torchvision==0.15.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# Streamlit and UI
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
requests==2.31.0
tqdm==4.66.1
shap==0.43.0

# Remove tensorflow and xgboost for now to reduce conflicts
# tensorflow>=2.6.0
# xgboost>=1.4.2 